DistributionTools.jl Documentation

This package is not registered, to install use Pkg.add("https://github.com/dfok/DistributionTools.jl)

DistributionTools provides some functions to work with a few multivariate distributions from Distributions.jl. The main additions are functions to obtain marginal and conditional distributions.

It also provides a function to obtain the cdf of a mixture of bivariate normals. For this GaussianDistributions.jl is used.

DistributionTools.jl Documentation

Obtaining marginal distributions

Obtaining conditional distributions

Other helper functions

Obtaining marginal distributions

DistributionTools.marginal — Function

marginal(d::MvNormal, ind::Int)

Return the marginal distribution of element ind from the multivariate normal distribution d

marginal(d::MvLogNormal, ind::Int)

Return the marginal distribution of element ind from the multivariate log normal distribution d

marginal(d::MvNormal, ind::Array{Int,1})

Return the marginal distribution of elements ind (vector) from the multivariate normal distribution d

marginal(d::MvLogNormal, ind::Array{Int,1})

Return the marginal distribution of elements ind (vector) from the multivariate log normal distribution d

marginal(d::MixtureModel, ind::Int)

Return the marginal distribution of element ind from a multivariate mixture distribution d

marginal(d::MixtureModel, ind::Array{Int,1})

Return the marginal distribution of elements ind (vector) from a multivariate mixture distribution d

marginal(d::MvTDist, ind::Int)

Return the marginal distribution of element ind from a multivariate t distribution d

marginal(d::MvTDist, ind::Array{Int,1})

Return the marginal distribution of elements ind (vector) from a multivariate t distribution d

Obtaining conditional distributions

DistributionTools.conditional — Function

conditional(d::MultivariateDistribution, cond::Int, xval::T) where T<:Real

Return conditional distribution when conditioning element cond of distribution d to value xval

conditional(d::MvLogNormal, cond::Array{Int,1}, xval::Array{<:Real,1}) where T<:Real

Return conditional distribution when conditioning element vector cond of LogNormal distribution d to value xval

conditional(d::MvNormal, cond::Array{Int,1}, xval::Array{<:Real,1}) where T<:Real

Returns conditional distribution when conditioning element vector cond of MvNormal distribution d to value xval

conditional(d::MvTDist, cond::Array{Int,1}, xval::Array{<:Real,1})

Obtain the conditional distribution of a multivariate T distribution

See Peng Ding (2016) "On the Conditional Distribution of the Multivariate t-Distribution", The American Statistician, 70:3, 293-295, DOI: 10.1080/00031305.2016.1164756

conditional(d::MixtureModel{Multivariate,Continuous,<:Distribution,<:DiscreteNonParametric}, cond::Array{Int,1}, xval::Array{<:Real,1})

Return the conditional distribution when conditioning element vector cond of Continuous Mixture distribution d to values xval

Drops elements from the mixture where the conditional mixture probability < eps()

Other helper functions

DistributionTools.cor2var — Function

cor2var(cor, var)

Transform a correlation + vector of variances to a 2x2 variance matrix

DistributionTools.var2cor — Function

var2cor(m)

Transform a variance matrix to a correlation matrix

DistributionTools.my_quantile — Function

my_quantile(d::UnivariateMixture{Continuous}, p::Real)

Return the p quantile of the univariate mixture distribution d

Adapted from quantile() from Distributions.jl, original has a (problem)[https://github.com/JuliaStats/Distributions.jl/issues/1869] with the bisect algorithm.

The initial interval does not contain solution, solved by ±0.000001

Distributions.cdf — Function

cdf(d::MixtureModel{Multivariate,Continuous, <:MvNormal, <:DiscreteNonParametric}, x::Vector{T}) where T<:Real

Return cdf of mixture of multivariate normals evaluated at x

Throws error when dimension of d ≠ 2